Moved get_testset_name to cmd_utils.py
